Refactored argparser for daac_data_subscriber.py into its own parser.py module
